backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
numpy>=1.26,<2.0
soundfile>=0.12.1
librosa>=0.10.2.post1
openai-whisper>=20231117
torch>=2.3.0
TTS>=0.22.0
transformers>=4.46.2,<4.50
sentencepiece>=0.1.99
python-multipart>=0.0.9
